more refactor and cleanup
