+++ package.json
@@ -2,7 +2,7 @@
     "name": "vscode-azurefunctions",
     "displayName": "Azure Functions",
     "description": "%azureFunctions.description%",
-    "version": "1.13.2-alpha.0",
+    "version": "1.13.4-alpha.0",
     "publisher": "ms-azuretools",
     "icon": "resources/azure-functions.png",
     "aiKey": "0c6ae279ed8443289764825290e4f9e2-1a736e7c-1324-4338-be46-fc2a58ae4d14-7255",
@@ -352,21 +352,6 @@
                 "category": "Azure Functions"
             },
             {
-                "command": "azureFunctions.createServiceConnector",
-                "title": "%azureFunctions.createServiceConnector%",
-                "category": "Azure Functions"
-            },
-            {
-                "command": "azureFunctions.deleteServiceConnector",
-                "title": "%azureFunctions.deleteServiceConnector%",
-                "category": "Azure Functions"
-            },
-            {
-                "command": "azureFunctions.validateServiceConnector",
-                "title": "%azureFunctions.validateServiceConnector%",
-                "category": "Azure Functions"
-            },
-            {
                 "command": "azureFunctions.eventGrid.sendMockRequest",
                 "title": "%azureFunctions.eventGrid.sendMockRequest%",
                 "category": "Azure Functions",
@@ -652,21 +637,6 @@
                     "command": "azureResourceGroups.refresh",
                     "when": "view =~ /(azureResourceGroups|azureFocusView)/ && viewItem =~ /azFunc.*folder/",
                     "group": "1@1"
-                },
-                {
-                    "command": "azureFunctions.createServiceConnector",
-                    "when": "view =~ /(azureResourceGroups|azureFocusView)/ && viewItem =~ /azFunc.*serviceConnectorGroupItem/",
-                    "group": "1@1"
-                },
-                {
-                    "command": "azureFunctions.deleteServiceConnector",
-                    "when": "view =~ /(azureResourceGroups|azureFocusView)/ && viewItem =~ /azFunc.*serviceConnectorItem/",
-                    "group": "1@1"
-                },
-                {
-                    "command": "azureFunctions.validateServiceConnector",
-                    "when": "view =~ /(azureResourceGroups|azureFocusView)/ && viewItem =~ /azFunc.*serviceConnectorItem/",
-                    "group": "1@2"
                 }
             ],
             "explorer/context": [
@@ -714,10 +684,6 @@
                 },
                 {
                     "command": "azureFunctions.viewProperties",
-                    "when": "never"
-                },
-                {
-                    "command": "azureFunctions.validateServiceConnector",
                     "when": "never"
                 }
             ],
@@ -1211,11 +1177,11 @@
         "@azure/core-client": "^1.7.3",
         "@azure/core-rest-pipeline": "^1.11.0",
         "@azure/storage-blob": "^12.5.0",
-        "@microsoft/vscode-azext-azureappservice": "^2.2.7",
+        "@microsoft/vscode-azext-azureappservice": "^3.1.0",
         "@microsoft/vscode-azext-azureappsettings": "^0.2.0",
-        "@microsoft/vscode-azext-azureutils": "^2.0.2",
+        "@microsoft/vscode-azext-azureutils": "^3.0.0",
         "@microsoft/vscode-azext-serviceconnector": "^0.1.3",
-        "@microsoft/vscode-azext-utils": "^2.1.0",
+        "@microsoft/vscode-azext-utils": "^2.3.1",
         "@microsoft/vscode-azureresources-api": "^2.0.4",
         "cross-fetch": "^4.0.0",
         "escape-string-regexp": "^4.0.0",
@@ -1234,5 +1200,12 @@
     },
     "extensionDependencies": [
         "ms-azuretools.vscode-azureresourcegroups"
-    ]
+    ],
+    "agentMetadata": {
+        "version": "1.0",
+        "getCommandsCommandId": "azureFunctions.agent.getCommands",
+        "runWizardCommandWithoutExecutionCommandId": "azureFunctions.agent.runWizardCommandWithoutExecution",
+        "runWizardCommandWithInputsCommandId": "azureFunctions.agent.runWizardCommandWithInputs",
+        "getAgentBenchmarkConfigsCommandId": "azureFunctions.agent.getAgentBenchmarkConfigs"
+    }
 }
