+++ package.nls.json
@@ -21,11 +21,9 @@
     "azureFunctions.createNewProjectWithDockerfile": "Create New Containerized Project...",
     "azureFunctions.createPythonVenv": "Create a virtual environment when creating a new Python project.",
     "azureFunctions.createSlot": "Create Slot...",
-    "azureFunctions.createServiceConnector": "Create connection...",
     "azureFunctions.deleteFunction": "Delete Function...",
     "azureFunctions.deleteFunctionApp": "Delete Function App...",
     "azureFunctions.deleteSlot": "Delete Slot...",
-    "azureFunctions.deleteServiceConnector": "Delete connection...",
     "azureFunctions.deploy": "Deploy to Function App...",
     "azureFunctions.deploySlot": "Deploy to Slot...",
     "azureFunctions.deploySubpath": "The default subpath of a workspace folder to use when deploying. If set, you will not be prompted for the folder path when deploying.",
@@ -104,7 +102,6 @@
     "azureFunctions.toggleAppSettingVisibility": "Toggle App Setting Visibility.",
     "azureFunctions.uninstallFuncCoreTools": "Uninstall Azure Functions Core Tools",
     "azureFunctions.validateFuncCoreTools": "Validate the Azure Functions Core Tools is installed before debugging.",
-    "azureFunctions.validateServiceConnector": "Validate",
     "azureFunctions.viewCommitInGitHub": "View Commit in GitHub",
     "azureFunctions.viewDeploymentLogs": "View Deployment Logs",
     "azureFunctions.viewProperties": "View Properties",
